fix project loading bug
