+++ packages/round-manager/src/features/api/utils.ts
@@ -1,6 +1,7 @@
 import { BigNumber, ethers } from "ethers";
 import {
   ApplicationMetadata,
+  GrantApplication,
   InputType,
   IPFSObject,
   RevisedMatch,
@@ -8,6 +9,7 @@ import {
 import { useEffect, useState } from "react";
 import Papa from "papaparse";
 import { getChainById } from "common";
+import { ProjectApplicationForManager } from "data-layer";
 
 export type SupportType = {
   name: string;
@@ -239,4 +241,53 @@ export const useMatchCSVParser = (file: File | null) => {
   }, [file]);
 
   return { data, loading, error };
+};
+
+export const convertApplications = (
+  dataLayerApplications: ProjectApplicationForManager[]
+) => {
+  const applications: GrantApplication[] = dataLayerApplications.flatMap(
+    (application) => {
+      if (application.canonicalProject === null) {
+        console.error(
+          `Canonical project not found for application ${application.id}`
+        );
+        return [];
+      }
+
+      return [
+        {
+          id: application.id,
+          applicationIndex: Number(application.id),
+          round: application.roundId,
+          status: application.status,
+          metadata: application.metadata,
+          project: {
+            ...application.metadata.application.project,
+            owners: application.canonicalProject.roles,
+            id: application.projectId,
+          },
+          projectId: application.projectId,
+          inReview: application.status === "IN_REVIEW",
+          recipient: application.metadata.application.recipient,
+          createdAt: "0",
+          projectsMetaPtr: { protocol: 1, pointer: "" },
+          payoutStrategy: {
+            strategyName: application.round.strategyName,
+            id: application.round.strategyAddress,
+            payouts: [],
+          },
+          distributionTransaction: application.distributionTransaction,
+          statusSnapshots: application.statusSnapshots.map((snapshot) => ({
+            ...snapshot,
+            updatedAt: new Date(snapshot.updatedAt),
+          })),
+          anchorAddress: application.anchorAddress,
+          answers: application.metadata.application.answers,
+        },
+      ];
+    }
+  );
+
+  return applications;
 };
